ci: refactor execution of Python tests
Previously we inconsistently used ubuntu-latest and ubuntu-20.04. We now
pin that to the most recent compatible version to avoid floating refs
(and the corresponding GitHub warning). As Python 3.6 and 3.7 is not
supported on Ubuntu 24.04, we still execute them on 20.04 (officially
EOL in April 2025).

To reduce the test code duplication, we carve out the pytest execution
into an action.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>